refactored my code
